app/requirements.txt
streamlit==1.28.0
wandb==0.16.0
langchain>=0.1.0
openai>=1.10.0
boto3==1.28.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.4
langchain-openai>=0.0.8
langchain-community==0.0.29
